+++ exchangelib/account.py
@@ -27,11 +27,28 @@ from .util import get_domain, peek
 log = getLogger(__name__)
 
 
+class Identity:
+    """Contains information that uniquely identifies an account. Currently only used for SOAP impersonation headers.
+    """
+    def __init__(self, primary_smtp_address=None, smtp_address=None, upn=None, sid=None):
+        """
+        :param primary_smtp_address: The primary email address associated with the account
+        :param smtp_address: The (non-)primary email address associated with the account
+        :param: upn: The User Principal Name (UPN) of this account
+        :param: sid: The security identifier (SID) of this account, in security descriptor definition language (SDDL)
+        form.
+        """
+        self.primary_smtp_address = primary_smtp_address
+        self.smtp_address = smtp_address
+        self.upn = upn
+        self.sid = sid
+
+
 class Account:
-    """Models an Exchange server user account. The primary key for an account is its PrimarySMTPAddress
+    """Models an Exchange server user account.
     """
     def __init__(self, primary_smtp_address, fullname=None, access_type=None, autodiscover=False, credentials=None,
-                 config=None, locale=None, default_timezone=None, sid=None, upn=None):
+                 config=None, locale=None, default_timezone=None):
         """
         :param primary_smtp_address: The primary email address associated with the account on the Exchange server
         :param fullname: The full name of the account. Optional.
@@ -43,9 +60,6 @@ class Account:
         :param locale: The locale of the user, e.g. 'en_US'. Defaults to the locale of the host, if available.
         :param default_timezone: EWS may return some datetime values without timezone information. In this case, we will
         assume values to be in the provided timezone. Defaults to the timezone of the host.
-        :param: sid: The security identifier (SID) of this account, in security descriptor definition language (SDDL)
-        form.
-        :param: upn: The User Principal Name (UPN) of this account
         """
         if '@' not in primary_smtp_address:
             raise ValueError("primary_smtp_address %r is not an email address" % primary_smtp_address)
@@ -83,23 +97,25 @@ class Account:
             self.ad_response, self.protocol = discover(
                 email=primary_smtp_address, credentials=credentials, auth_type=auth_type, retry_policy=retry_policy
             )
-            self.primary_smtp_address = self.ad_response.autodiscover_smtp_address
+            primary_smtp_address = self.ad_response.autodiscover_smtp_address
         else:
             if not config:
                 raise AttributeError('non-autodiscover requires a config')
-            self.primary_smtp_address = primary_smtp_address
+            primary_smtp_address = primary_smtp_address
             self.ad_response = None
             self.protocol = Protocol(config=config)
 
-        # Other ways of identifying the account. Currently only used for SOAP impersonation headers.
-        self.sid = sid
-        self.upn = upn
-        self.smtp_address = None
+        # Other ways of identifying the account can be added later
+        self.identity = Identity(primary_smtp_address=primary_smtp_address)
 
         # We may need to override the default server version on a per-account basis because Microsoft may report one
         # server version up-front but delegate account requests to an older backend server.
         self.version = self.protocol.version
         log.debug('Added account: %s', self)
+
+    @property
+    def primary_smtp_address(self):
+        return self.identity.primary_smtp_address
 
     @threaded_cached_property
     def admin_audit_logs(self):